pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "audio-metadata"
description = "A library for reading and, in the future, writing metadata from audio files."
version = "0.1.0"

license = "MIT"

authors = ["thebigmunch <mail@thebigmunch.me>"]

readme = "README.md"

repository = "https://github.com/thebigmunch/audio-metadata"
homepage = "https://github.com/thebigmunch/audio-metadata"

[tool.poetry.dependencies]
python = "^3.6"

attrs = "^18.2"
bidict = "^0.17"
more-itertools = "^4.0"
pprintpp = "^0.4"

flake8 = { version = "^3.5", optional = true }
flake8-builtins = { version = "^1.0", optional = true }
flake8-import-order = { version = "^0.18", optional = true }
flake8-import-order-tbm = { version = "^1.0.0", optional = true }
sphinx = { version= "^1.8", optional = true}

[tool.poetry.extras]
dev = [ "flake8", "flake8-builtins", "flake8-import-order", "flake8-import-order-tbm", "sphinx" ]
docs = [ "sphinx" ]
lint = [ "flake8", "flake8-builtins", "flake8-import-order", "flake8-import-order-tbm" ]
